updates mysqlprovider Get methods with explicit block for sql.ErrNoRows
the purpose of this change is to reserve the error path only for
network-level errors. The sql package treats a sql.ErrNoRows as an
error (which i do not want in the error path) so i am explicitly
checking for this error type so that errors of this
type do not escape the mysqlprovider driver layer (and having an
explicit code block gives an indication of whether this rule is checked
in test coverage)

since sql.ErrNoRows does not escape the driver layer, what is returned
instead is a zero-value struct of the return type. so callers of this
driver may need to check the return value if they need to determine
whether there was an actual value found for the query or not. this still
requries introspection and doesnt really save any work compared to
returning sql.ErrNoRows, however, it is semantically distinct
because i am returning on the "success" path rather than the "error"
path and the driver remains abstract in the sense that no reference to
"sql" escapes its layer.