updates AuthUser tests to fix issue with mock return value
in e3b05, the change caused the usecases AuthUser tests to fail because
it calls provider GetUser mock which was not returning any values, which
lead to a nil pointer reference error. this update adds a return value
for provider GetUser so that calls to usecases GetUser does not error.