custom change to vendor/github.com/grpc-ecosystem/grpc-gateway/runtime
this changes the defaultMarshaler in the runtime so that it uses the
canonical camelCasing for emitting json values from the gateway.
Previously, the json properties were being returned as snake_case.

This change should be present in the next official release of
grpc-ecosystem/grpc-gateway, but that has not been released yet. So this
custom modification of the vendor/ package is a temporary hack.