replace memcached Conn with Cacher service in authentication service
this changes authentication service so it requires a Cacher service to
instantiate instead of a memcached connection. this change allows the
authentication service to depend upon an abstraction instead of a
concrete instance of memcached. Effectively, the Cacher service is a
driver for the memcached connection but with a much smaller api.