add ability to perform db migrations based on DB_SCHEMA_VERSION env var
this change adds logic to server.go that checks for the existance of
DB_SCHEMA_VERSION env var, and if it exists and is a different value
than the currently installed version in the database, then it will try
to perform the migration. this process is considered part of the
server.go startup process, and therefore, if anything goes wrong it
constitutes a panic.

also modified the server Dockerfile to use golang:1.9 (required by
golang-migrate/migrate/database/mysql) and copies migrations/ into the
server container so that the server can access the migration files
locally when it starts up.

also removes db service Dockerfile and changes it to image mysql:5
instead of a build in the docker-compose. The db service was originally
using a Dockerfile so that it could copy over the docker entrypoint
files which contain the version 1 of the database schema. However, since
this commit implements a migration system, it means that we no longer
need to define a docker entrypoint files. The migration system will
automatically install version 1 of the database schema on server startup
if the database is being used for the first time.